actual .zip for the whole mod
I can't remember but you might have to edit the actual player file to force them to equip the manyjump tech. Either it crashes the game when you try to equip it or you simply can't have it unequipped. Doesn't really matter to me and I doubt it matters to anyone else, because who is actually looking at this? I'd totally ask for player feedback and try to bug fix, but I don't play the game frequently anymore, sorry :( . Will properly fix if asked tho :)